refactor: refactor imports on tests
