+++ src/shared/api/index.ts
@@ -1,2 +1,2 @@
-export { checkOneTimePassword } from './supabase/check-otp'
-export { signUpWithMailAndPassword } from './supabase/sign-up-with-mail-and-password'
+export { checkOneTimePassword } from './supabase/auth/check-otp'
+export { signUpWithMailAndPassword } from './supabase/auth/sign-up-with-mail-and-password'
