+++ src/shared/api/supabase/auth/sign-up-with-mail-and-password.ts
